rename mysql helper file to be more generic
