+++ setup_test_db.py
@@ -1,5 +1,5 @@
 import lighthouse.config.test as config  # type: ignore
-from lighthouse.helpers.warehouses_db import create_mysql_connection_engine
+from lighthouse.helpers.mysql_db import create_mysql_connection_engine
 
 # Set up a basic MLWH db for testing
 """Drop and recreate required tables."""
