+++ tests/conftest.py
@@ -26,7 +26,7 @@ from .data.fixture_data import (
     MLWH_SAMPLE_STOCK_RESOURCE
 )
 
-from lighthouse.helpers.warehouses_db import create_mysql_connection_engine, get_table
+from lighthouse.helpers.mysql_db import create_mysql_connection_engine, get_table
 
 
 @pytest.fixture
